# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "jwt token test",
        "title": "Swagger Example",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/first": {
            "post": {
                "description": "sdf",
                "produces": [
                    "application/json"
                ],
                "summary": "get acc and ref tokens",
                "operationId": "getFirst",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/second": {
            "post": {
                "description": "sfdgddf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "refresh acc and ref tokens",
                "operationId": "Second",
                "parameters": [
                    {
                        "description": "token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hand.tokenJs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "hand.tokenJs": {
            "type": "object",
            "properties": {
                "AccToken": {
                    "type": "string"
                },
                "RefToken": {
                    "type": "string"
                }
            }
        }
    }
}